requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
flask==3.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1